+++ src/Organisms/ImageUpload/upload.css
@@ -1,0 +1,51 @@
+.text-center {
+  display: flex;
+  height: 100%;
+  justify-content: center;
+  align-content: center;
+  text-align: center;
+}
+
+/* Images */
+.file-list {
+  margin: 20px 0;
+  padding: 0;
+  list-style: none;
+  display: flex;
+  flex-wrap: wrap;
+}
+.file-item {
+  width: calc(33% - 18px);
+  height: auto;
+  border: 1px solid #eaeaea;
+  margin: 8px;
+  cursor: move;
+  transition: all 0.2s linear;
+}
+.file-img {
+  width: 100%;
+  height: 100%;
+}
+
+/* Tablets */
+@media (max-width: 1000px) {
+  .App {
+    max-width: 600px;
+  }
+  .file-item {
+    width: calc(50% - 22px);
+    height: 200px;
+  }
+}
+
+/* Mobiles */
+@media (max-width: 640px) {
+  .App {
+    max-width: 100%;
+    padding: 0 15px;
+  }
+  .file-item {
+    width: calc(100% - 22px);
+    height: 200px;
+  }
+}
